split writer and 'split' module off
