+++ examples/Cargo.toml
@@ -4,7 +4,6 @@ version = "0.0.0"
 edition = "2018"
 publish = false
 
-
 [dev-dependencies]
 twitchchat = { version = "0.14", path = "../twitchchat", features = ["async", "writer"] }
 
@@ -17,7 +16,7 @@ twitchchat_tokio02   = { version = "0.1", path = "../twitchchat_tokio02" }
 
 anyhow         = "1.0.33"
 async-executor = "1.3.0"
-futures-lite = "1.11.1"
+futures-lite   = "1.11.1"
 
 # the various runtimes
 tokio     = { version = "0.3.0", features = ["full", "macros"] }
