+++ twitchchat/src/lib.rs
@@ -58,16 +58,6 @@ For just encoding messages:
 * [encoder]
 */
 
-macro_rules! cfg_async {
-    ($($item:item)*) => {
-        $(
-            #[cfg(feature = "async")]
-            #[cfg_attr(docsrs, doc(cfg(feature = "async")))]
-            $item
-        )*
-    };
-}
-
 /// The Twitch IRC address for non-TLS connections
 pub const TWITCH_IRC_ADDRESS: &str = "irc.chat.twitch.tv:6667";
 
@@ -96,147 +86,15 @@ mod macros;
 
 pub mod decoder;
 pub use decoder::{DecodeError, Decoder};
-cfg_async! {
-    pub use decoder::AsyncDecoder;
-    /// A boxed `AsyncRead` trait
-    pub type BoxedRead = Box<dyn futures_lite::AsyncRead + Send + Unpin>;
-    /// A boxed `AsyncDecoder`
-    pub type BoxedAsyncDecoder = AsyncDecoder<BoxedRead>;
-}
+cfg_async! { pub use decoder::AsyncDecoder; }
 
 pub mod encoder;
 pub use encoder::Encoder;
-cfg_async! {
-    pub use encoder::AsyncEncoder;
-    /// A boxed `BoxedWrite` trait
-    pub type BoxedWrite = Box<dyn futures_lite::AsyncWrite + Send + Unpin>;
-    /// A boxed `AsyncEncoder`
-    pub type BoxedAsyncEncoder = AsyncEncoder<BoxedWrite>;
-}
+cfg_async! { pub use encoder::AsyncEncoder; }
 
-cfg_async! {
-/// Split an Async IO object and return Decoder/Encoder
-pub fn make_pair<IO>(io: IO) -> (AsyncDecoder<futures_lite::io::ReadHalf< IO>>, AsyncEncoder<futures_lite::io::WriteHalf<IO>>)
-where
-    IO: futures_lite::AsyncRead + futures_lite::AsyncWrite,
-    IO: Send + Unpin + 'static,
-{
-    let (read, write) = futures_lite::io::split(io);
-    (AsyncDecoder::new(read), AsyncEncoder::new(write))
-}
-}
+pub mod split;
 
-cfg_async! {
-/// Split an Async IO object and return type-erased Decoder/Encoder
-pub fn make_boxed_pair<IO>(io: IO) -> (BoxedAsyncDecoder, BoxedAsyncEncoder)
-where
-    IO: futures_lite::AsyncRead + futures_lite::AsyncWrite,
-    IO: Send + Unpin + 'static,
-{
-    let (read, write) = futures_lite::io::split(io);
-    let read: BoxedRead = Box::new(read);
-    let write: BoxedWrite = Box::new(write);
-    (AsyncDecoder::new(read), AsyncEncoder::new(write))
-}
-}
-
-#[cfg(feature = "writer")]
-#[allow(missing_docs)]
-pub mod writer {
-    use super::*;
-
-    cfg_async! { use futures_lite::io::AsyncWrite; }
-
-    use std::io::Write;
-
-    enum Packet {
-        Data(Box<[u8]>),
-        Quit,
-    }
-
-    #[derive(Clone)]
-    pub struct MpscWriter {
-        tx: flume::Sender<Packet>,
-        wait_for_it: flume::Receiver<()>,
-    }
-
-    impl std::fmt::Debug for MpscWriter {
-        fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-            f.debug_struct("MpscWriter").finish()
-        }
-    }
-
-    impl MpscWriter {
-        pub fn from_encoder<W>(encoder: Encoder<W>) -> Self
-        where
-            W: Write + Send + 'static,
-        {
-            let mut encoder = encoder;
-            let (tx, rx) = flume::unbounded();
-            let (stop_tx, wait_for_it) = flume::bounded(1);
-            Self::_start(rx, stop_tx, move |data| encoder.encode(data));
-            Self { tx, wait_for_it }
-        }
-
-        cfg_async! {
-        pub fn from_async_encoder<W>(encoder: AsyncEncoder<W>) -> Self
-        where
-            W: AsyncWrite + Send + Unpin + 'static,
-        {
-            let mut encoder = encoder;
-            use futures_lite::future::block_on;
-            let (tx, rx) = flume::unbounded();
-            let (stop_tx, wait_for_it) = flume::bounded(1);
-            Self::_start(rx, stop_tx, move |data| block_on(encoder.encode(data)));
-            Self { tx, wait_for_it }
-        }
-        }
-
-        fn _start<F>(receiver: flume::Receiver<Packet>, stop: flume::Sender<()>, mut encode: F)
-        where
-            F: FnMut(Box<[u8]>) -> std::io::Result<()>,
-            F: Send + 'static,
-        {
-            // TODO don't do this for the async one
-            let _ = std::thread::spawn(move || {
-                let _stop = stop;
-                for msg in receiver {
-                    match msg {
-                        Packet::Data(data) => encode(data)?,
-                        Packet::Quit => break,
-                    }
-                }
-                std::io::Result::Ok(())
-            });
-        }
-
-        pub fn send(&self, enc: impl Encodable) -> std::io::Result<()> {
-            let mut buf = Vec::new();
-            enc.encode(&mut buf)?;
-
-            if let Err(flume::TrySendError::Disconnected(..)) =
-                self.tx.try_send(Packet::Data(buf.into_boxed_slice()))
-            {
-                return Err(std::io::ErrorKind::BrokenPipe.into());
-            }
-            Ok(())
-        }
-
-        cfg_async! {
-        pub async fn shutdown(self) {
-            let _ = self.send(crate::commands::raw("QUIT\r\n"));
-            let _ = self.tx.try_send(Packet::Quit);
-            let _ = self.wait_for_it.into_recv_async().await;
-        }
-        }
-
-        pub fn shutdown_sync(self) {
-            let _ = self.send(crate::commands::raw("QUIT\r\n"));
-            let _ = self.tx.try_send(Packet::Quit);
-            let _ = self.wait_for_it.recv();
-        }
-    }
-}
+cfg_writer! { pub mod writer; }
 
 pub mod runner;
 pub use runner::Error;
