+++ twitchchat/src/runner/mod.rs
@@ -22,11 +22,13 @@ cfg_async! {
     pub use timeout::idle_detection_loop;
     pub use wait_for_ready::wait_until_ready;
 
-    #[cfg(feature = "writer")]
-    pub use timeout::respond_to_idle_events;
+    cfg_writer! {
+        pub use timeout::respond_to_idle_events;
+    }
 
     pub use wait_for::wait_for;
 
-    #[cfg(feature = "ws")]
-    pub mod ws;
+    cfg_ws! {
+        pub mod ws;
+    }
 }
